modularize and refactor strategies for scalability
