+++ .env.development
@@ -4,9 +4,9 @@ PORT=3000
 # Database Connection
 DB_HOST=localhost
 DB_PORT=5432
-DB_USER=lagodxy
-DB_PASSWORD=4633922
-DB_NAME=lyricsflip_store
+DB_USER=postgres
+DB_PASSWORD=1234
+DB_NAME=streller_db
 
 # Authentication
 JWT_SECRET=mysecretJourney
