+++ src/auth/auth.controller.ts
@@ -13,7 +13,8 @@ import {
   BadRequestException,
 } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { AuthService } from './auth.service';
+// import { AuthService } from './auth.service';
+
 import { LoginDto } from './dto/login.dto';
 import { RegisterDto } from './dto/register.dto';
 import { ResetPasswordDto } from './dto/reset-password.dto';
@@ -25,12 +26,13 @@ import { ApiBearerAuth, ApiTags, ApiOperation, ApiResponse } from '@nestjs/swagg
 import { UsersService } from 'src/users/services/users.service';
 import { RateLimitGuard } from 'src/common/guards/rate-limiter.guard';
 import { Strategy as JwtStrategy, ExtractJwt } from 'passport-jwt';
+import { JwtLocalStrategy } from './strategies/jwt-local.strategy';
 
 @ApiTags('authentication')
 @Controller('auth')
 export class AuthController {
   constructor(
-    private readonly authService: AuthService,
+    private readonly jwtLocalStrategy: JwtLocalStrategy,
     private readonly usersService: UsersService,
     private readonly passwordValidationService: PasswordValidationService,
   ) {}
@@ -42,18 +44,18 @@ export class AuthController {
   @ApiResponse({ status: 401, description: 'Invalid credentials' })
   @HttpCode(HttpStatus.OK)
   async login(@Body() body: LoginDto) {
-    const user = await this.authService.validateUser(body.email, body.password);
+    const user = await this.jwtLocalStrategy.validateUser(body.email, body.password);
     if (!user) {
       throw new UnauthorizedException('Invalid credentials');
     }
-    return this.authService.login(user);
+    return this.jwtLocalStrategy.login(user);
   }
 
   @Post('register')
   async register(@Body() registerDto: RegisterDto) {
     try {
       const { email, password, ...userData } = registerDto;
-      return await this.authService.register(email, password, userData);
+      return await this.jwtLocalStrategy.register(registerDto);
     } catch (error) {
       if (error instanceof BadRequestException) {
         throw error;
@@ -78,23 +80,23 @@ export class AuthController {
     if (!body.userId || !body.refreshToken) {
       throw new UnauthorizedException('Missing credentials');
     }
-    return this.authService.refreshToken(body.userId, body.refreshToken);
+    return this.jwtLocalStrategy.refreshToken(body.userId, body.refreshToken);
   }
 
   @Post('forgot-password')
   async forgotPassword(@Body('email') email: string) {
-    return this.authService.requestPasswordReset(email);
+    return this.jwtLocalStrategy.requestPasswordReset(email);
   }
 
   @Get('validate-reset-token')
   async validateToken(@Query('token') token: string) {
-    await this.authService.validateResetToken(token);
+    await this.jwtLocalStrategy.validateResetToken(token);
     return { message: 'Token is valid' };
   }
 
   @Post('reset-password')
   async resetPassword(@Body() resetDto: ResetPasswordDto) {
-    return this.authService.resetPassword(resetDto.token, resetDto.newPassword);
+    return this.jwtLocalStrategy.resetPassword(resetDto.token, resetDto.newPassword);
   }
 
   @UseGuards(JwtAuthGuard)
@@ -110,7 +112,7 @@ export class AuthController {
       throw new BadRequestException('Current password and new password are required');
     }
 
-    return this.authService.changePassword(userId, currentPassword, newPassword);
+    return this.jwtLocalStrategy.changePassword(userId, currentPassword, newPassword);
   }
 
   @Post('logout')
@@ -120,7 +122,7 @@ export class AuthController {
   @ApiResponse({ status: 200, description: 'Logout successful' })
   @HttpCode(HttpStatus.OK)
   async logout(@Request() req) {
-    await this.authService.logout(req.user.userId);
+    await this.jwtLocalStrategy.logout(req.user.userId);
     return { message: 'Logout successful' };
   }
 }
