+++ src/auth/strategies/auth-strategy.interface.ts
@@ -1,0 +1,6 @@
+export interface IAuthStrategy {
+  name: string;
+  validate(credentials: any): Promise<any>;
+  login(user: any): Promise<any>;
+  register?(credentials: any): Promise<any>;
+}
